To build an ATM Interface using Python programming language.
The ATMs in our cities are built on Python, as we have seen them. It is a console-based application with five different classes. In order to use the system, the user must enter his/her user ID and pin when it starts. Once the details are entered successfully, ATM functionality is unlocked.